RatAttack: fixed RatAttack.parseRatsignal
VoicaAttack can handle control characters within a commond definition.
Sadly it errors out if you try to _import_ a profile that has said
control characters in them …

The workaround is to set the delimiter variable (`^B`) in an inline
function. This way the character itself does not appear in the exported
profile and everything works fine.