EliteAttack: moved logic from `Jumped` to `FSD engaged`
Getting the EDSM body count and outdated stations will now happen on
engaging FSD instead of finishing the jump.

With the increased latency on the EDSM API I’m seeing lately this was
very much needed to get the body count in time for the discovery scan
after the jump.

fixes #85